# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: OneQR Business Growth Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List user's businesses
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: List of businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"
    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Create a new business
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{id}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
    patch:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update business
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessRequest"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /b/{slug}:
    get:
      operationId: getPublicProfile
      tags: [businesses]
      summary: Get public business profile (QR destination)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Public business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicBusinessProfile"
  /businesses/{id}/links:
    get:
      operationId: listLinks
      tags: [links]
      summary: List social links
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialLink"
    post:
      operationId: createLink
      tags: [links]
      summary: Add a social link
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLinkRequest"
      responses:
        "201":
          description: Created link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialLink"
  /businesses/{id}/qrcodes:
    get:
      operationId: listQRCodes
      tags: [qrcodes]
      summary: List QR codes
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of QR codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QRCode"
  /businesses/{id}/offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create offer
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfferRequest"
      responses:
        "201":
          description: Created offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Business:
      type: object
      properties:
        id: { type: string }
        ownerId: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        category: { type: string }
        logoUrl: { type: string, nullable: true }
        coverImageUrl: { type: string, nullable: true }
        primaryColor: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        website: { type: string, nullable: true }
        address: { type: string, nullable: true }
        googleMapsUrl: { type: string, nullable: true }
        googleReviewUrl: { type: string, nullable: true }
        whatsappNumber: { type: string, nullable: true }
        isActive: { type: boolean }
        isVerified: { type: boolean }
        totalScans: { type: integer }
        totalClicks: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, ownerId, name, slug, category, primaryColor, isActive, isVerified, totalScans, totalClicks, createdAt]
    CreateBusinessRequest:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        category: { type: string }
        phone: { type: string }
        website: { type: string }
        address: { type: string }
        googleMapsUrl: { type: string }
        googleReviewUrl: { type: string }
        whatsappNumber: { type: string }
      required: [name, slug, category]
    UpdateBusinessRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        category: { type: string }
        primaryColor: { type: string }
        phone: { type: string }
        website: { type: string }
        address: { type: string }
        googleMapsUrl: { type: string }
        googleReviewUrl: { type: string }
        whatsappNumber: { type: string }
        isActive: { type: boolean }
    PublicBusinessProfile:
      type: object
      properties:
        business:
          $ref: "#/components/schemas/Business"
        links:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
        offers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
      required: [business, links, offers]
    SocialLink:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        platform: { type: string }
        label: { type: string, nullable: true }
        url: { type: string }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        clicks: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, businessId, platform, url, sortOrder, isActive, clicks, createdAt]
    CreateLinkRequest:
      type: object
      properties:
        platform: { type: string }
        label: { type: string }
        url: { type: string }
        sortOrder: { type: integer }
      required: [platform, url]
    QRCode:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        label: { type: string }
        destinationUrl: { type: string }
        qrImageUrl: { type: string, nullable: true }
        isActive: { type: boolean }
        scans: { type: integer }
        createdAt: { type: string, format: date-time }
      required: [id, businessId, label, destinationUrl, isActive, scans, createdAt]
    Offer:
      type: object
      properties:
        id: { type: string }
        businessId: { type: string }
        title: { type: string }
        description: { type: string, nullable: true }
        emoji: { type: string }
        isActive: { type: boolean }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, businessId, title, emoji, isActive, createdAt]
    CreateOfferRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        emoji: { type: string }
        expiresAt: { type: string, format: date-time }
      required: [title]
...